Structure PairingUI into seprate files
