+++ ENGAGEHF/Devices/Model/DevicePairingError.swift
