+++ ENGAGEHF/Devices/Model/HealthDevice.swift
